+++ src/commands/transaction/reconstruct_transaction/mod.rs
@@ -234,8 +234,8 @@ fn action_transformation(
             )))
         }
         Action::Pledge(pledge_action) => {
-                Ok(Some(add_action::CliActionSubcommand::Stake(
-                add_action::stake::CliPledgeAction {
+                Ok(Some(add_action::CliActionSubcommand::Pledge(
+                add_action::pledge::CliPledgeAction {
                     pledge_amount: Some(crate::types::unc_token::UncToken::from_yoctounc(pledge_action.pledge)),
                     public_key: Some(pledge_action.public_key.into()),
                     next_action: None